move JS imports and exports to separate projects
