+++ cli/.js/src/main/scala/aqua/ipfs/js/IpfsApi.scala
@@ -1,7 +1,5 @@
 package aqua.ipfs.js
 
-import aqua.js.FluencePeer
-
 import scala.scalajs.js
 import scala.scalajs.js.annotation.{JSExportAll, JSImport}
